+++ modules/versions.lua
@@ -10,7 +10,7 @@ return {
   ["Gen1ItemInfo"] = "maintained",
   ["Gen1MenuManager"] = "0.4.0",
   ["Gen1ModMenu"] = "0.9.0",
-  ["Gen1ModernBag"] = "1.13.0",
+  ["Gen1ModernBag"] = "1.13.1",
   ["Gen1Party"] = "1.8.1",
   ["WidescreenBattleIntro"] = "maintained",
 }